check registration
