+++ src/services/auth.service.ts
@@ -21,3 +21,28 @@ export const logout = () => {
 export const getCurrentUser = () => {
   return JSON.parse(localStorage.getItem("user") as string);
 }
+
+
+export const registration = async (  
+  userName: string,
+  firstName: string,
+  lastName: string,
+  email: string,
+  password: string,
+  confirmPassword: string,) => {
+  return await 
+    axios.post("/registration", {
+      userName,
+      firstName,
+      lastName,
+      email,
+      password,
+      confirmPassword
+    })
+    .then((res) => {
+      if (res.data.token) {
+        localStorage.setItem("newUser", JSON.stringify(res.data.token));
+      }
+      return res.data;
+    });
+};
